move `WINDOW` to ./constants.ts
