+++ packages/replay/src/constants.ts
@@ -1,3 +1,10 @@
+import { GLOBAL_OBJ } from '@sentry/utils';
+
+// exporting WINDOW from within the @sentry/replay instead of importing it from @sentry/browser
+// this avoids the Browser package being bundled into the CDN bundle as well as a
+// circular dependency between the Browser and Replay packages in the future
+export const WINDOW = GLOBAL_OBJ as typeof GLOBAL_OBJ & Window;
+
 export const REPLAY_SESSION_KEY = 'sentryReplaySession';
 export const REPLAY_EVENT_NAME = 'replay_event';
 export const RECORDING_EVENT_NAME = 'replay_recording';
