+++ packages/replay/src/session/saveSession.ts
@@ -1,5 +1,4 @@
-import { WINDOW } from '../types';
-import { REPLAY_SESSION_KEY } from './constants';
+import { REPLAY_SESSION_KEY, WINDOW } from '../constants';
 import { Session } from './Session';
 
 export function saveSession(session: Session): void {
