+++ packages/replay/test/unit/index-handleGlobalEvent.test.ts
@@ -2,8 +2,8 @@ import { Error } from '@test/fixtures/error';
 import { Transaction } from '@test/fixtures/transaction';
 import { resetSdkMock } from '@test/mocks';
 
+import { REPLAY_EVENT_NAME } from '../../src/constants';
 import { Replay } from './../../src';
-import { REPLAY_EVENT_NAME } from './../../src/session/constants';
 import { useFakeTimers } from './../utils/use-fake-timers';
 
 useFakeTimers();
