+++ packages/replay/test/unit/session/fetchSession.test.ts
@@ -1,6 +1,5 @@
-import { REPLAY_SESSION_KEY } from '../../../src/session/constants';
+import { REPLAY_SESSION_KEY, WINDOW } from '../../../src/constants';
 import { fetchSession } from '../../../src/session/fetchSession';
-import { WINDOW } from '../../../src/types';
 
 const oldSessionStorage = WINDOW.sessionStorage;
