+++ packages/replay/test/unit/session/saveSession.test.ts
@@ -1,7 +1,6 @@
-import { REPLAY_SESSION_KEY } from '../../../src/session/constants';
+import { REPLAY_SESSION_KEY, WINDOW } from '../../../src/constants';
 import { saveSession } from '../../../src/session/saveSession';
 import { Session } from '../../../src/session/Session';
-import { WINDOW } from '../../../src/types';
 
 beforeAll(() => {
   WINDOW.sessionStorage.clear();
