+++ packages/replay/test/unit/stop.test.ts
@@ -2,9 +2,8 @@ import * as SentryUtils from '@sentry/utils';
 // mock functions need to be imported first
 import { BASE_TIMESTAMP, mockRrweb, mockSdk } from '@test';
 
-import { WINDOW } from '../../src/types';
+import { SESSION_IDLE_DURATION, WINDOW } from '../../src/constants';
 import { Replay } from './../../src';
-import { SESSION_IDLE_DURATION } from './../../src/session/constants';
 import { useFakeTimers } from './../utils/use-fake-timers';
 
 useFakeTimers();
